feat: Add initial configuration files for prettier and robots.txt
This commit adds the initial configuration files for prettier and robots.txt. The `.prettierrc` file is added to the `CRUD functions\pharmacy-backend` directory and the `pharmacy-backend` directory is deleted. Similarly, the `robots.txt` file is added to the `CRUD functions\pharmacy-frontend\public` directory and the `pharmacy-frontend/public` directory is deleted.